Use a symbolic link for the pre-commit client-side hook

+++ autogen.sh
@@ -23,7 +23,7 @@ fi
 . common/gst-autogen.sh
 
 # install pre-commit hook for doing clean commits
-if test ! -x .git/hooks/pre-commit;
+if test ! \( -x .git/hooks/pre-commit -a -L .git/hooks/pre-commit \);
 then
     cp common/hooks/pre-commit.hook .git/hooks/pre-commit
     chmod +x .git/hooks/pre-commit
